requirements.txt
appdirs==1.4.4
bcrypt==3.1.7
cffi==1.16.0
cryptography==2.8
Fabric3==1.14.post1
paramiko==2.12.0
pip==22.0.2
pycparser==2.21
PyNaCl==1.3.0
pyparsing==3.1.1
setuptools==59.6.0
six==1.16.0

